friend.html, AccountController 수정

+++ src/main/java/meetmeet/model/dao/FriendRequestRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import meetmeet.model.entity.FriendRequest;
 
-public interface FriendRequestRepository extends CrudRepository<FriendRequest, Long> {
+public interface FriendRequestRepository extends CrudRepository<FriendRequest, Integer> {
 	@Query("SELECT fr.requestId.accountId, fr.requestedId.accountId FROM FriendRequest fr WHERE fr.requestId.accountId=:requestId")
 	public abstract List<List<String>> findFriendRequestByRequestId(String requestId);
